기타 수정-Entity에서 DTO를 모르도록 수정
- Entity에서 DTO를 모르도록 수정
  - 확장함수를 DTO 쪽으로 빼냄
- TODO 엔티티가 커스텀한 exception을 알고 있는 게 괜찮을지 고민해보기
